+++ tests/test_principals.sh
@@ -1,31 +1,6 @@
 #!/bin/bash
 # shellcheck disable=SC2128
 
-# DEPRECATED WAY TO ADD PRINCIPALS
-RESP=$(curl -s -X PATCH -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}" -d "principals=${GUEST_B_USERNAME},test")
-if [[ "${RESP}" == *"OK: principals=${GUEST_B_USERNAME},test for ${GUEST_B_USERNAME}"* ]]; then
-    echo "[OK][DEPRECATED] Test add principal 'test' to ${GUEST_B_USERNAME}"
-else
-    echo "[FAIL][DEPRECATED] Test add principal 'test' to ${GUEST_B_USERNAME} : ${RESP}"
-fi
-
-RESP=$(curl -s -X PATCH -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}" -d "principals=${GUEST_B_USERNAME},test-with-dash")
-if [[ "${RESP}" == *"OK: principals=${GUEST_B_USERNAME},test-with-dash for ${GUEST_B_USERNAME}"* ]]; then
-    echo "[OK][DEPRECATED] Test add principal 'test-with-dash' to ${GUEST_B_USERNAME}"
-else
-    echo "[FAIL][DEPRECATED] Test add principal 'test-with-dash' to ${GUEST_B_USERNAME} : ${RESP}"
-fi
-
-RESP=$(curl -s -X PATCH -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}" -d "principals=${GUEST_B_USERNAME},${BADTEXT}")
-if [[ "${RESP}" == *"Error: invalid principals."* ]]; then
-    echo "[OK][DEPRECATED] Test add wrong principal '${BADTEXT}' to ${GUEST_B_USERNAME}"
-else
-    echo "[FAIL][DEPRECATED] Test add wrong principal '${BADTEXT}' to ${GUEST_B_USERNAME} : ${RESP}"
-fi
-
-
-
-# NEW WAY TO ADD PRINCIPALS
 RESP=$(curl -s -X POST "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "add=test-single")
 if [ "${RESP}" == "Error: No realname option given." ]; then
     echo "[OK] Test add principal without realname,password"
@@ -65,21 +40,32 @@ fi
 ## ADMIN ADD PRINCIPAL GUEST B ##
 #################################
 RESP=$(curl -s -X POST -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "add=test-single")
-if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-with-dash,test-single'" ]; then
+if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-single'" ]; then
     echo "[OK] Test add principal 'test-single' to ${GUEST_B_USERNAME}"
 else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test add principal 'test-single' to ${GUEST_B_USERNAME} : ${RESP}"
 fi
 
+RESP=$(curl -s -X POST -d "username=${GUEST_B_USERNAME}&realname=${GUEST_B_REALNAME}&password=${GUEST_B_PASSWORD}&pubkey=${GUEST_B_PUB_KEY}" "${CASSH_SERVER_URL}"/client)
+echo "${RESP}" > /tmp/test-cert
+OUTPUT=$(echo $(echo -n "$(ssh-keygen -L -f /tmp/test-cert 2>&1)"))
+if [[ "${OUTPUT}" == *"Principals: ${GUEST_B_USERNAME} test-single Critical"* ]]; then
+    echo "[OK] Test signing key with updated principals"
+else
+    echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test signing key with updated principals: ${OUTPUT}"
+fi
+rm -f /tmp/test-cert
+
+
 RESP=$(curl -s -X POST -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "add=test-single")
-if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-with-dash,test-single'" ]; then
+if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-single'" ]; then
     echo "[OK] Test add duplicate principal 'test-single' to ${GUEST_B_USERNAME}"
 else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test add duplicate principal 'test-single' to ${GUEST_B_USERNAME} : ${RESP}"
 fi
 
 RESP=$(curl -s -X POST -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "remove=test-single")
-if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-with-dash'" ]; then
+if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME}'" ]; then
     echo "[OK] Test remove principal 'test-single' to ${GUEST_B_USERNAME} which doesn't exists"
 else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test remove principal 'test-single' to ${GUEST_B_USERNAME} which doesn't exists : ${RESP}"
@@ -89,11 +75,21 @@ fi
 ## ADMIN REMOVE PRINCIPAL GUEST B ##
 ####################################
 RESP=$(curl -s -X POST -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "remove=test-single")
-if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-with-dash'" ]; then
+if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME}'" ]; then
     echo "[OK] Test remove principal 'test-single' to ${GUEST_B_USERNAME}"
 else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test remove principal 'test-single' to ${GUEST_B_USERNAME} : ${RESP}"
 fi
+
+RESP=$(curl -s -X POST -d "username=${GUEST_B_USERNAME}&realname=${GUEST_B_REALNAME}&password=${GUEST_B_PASSWORD}&pubkey=${GUEST_B_PUB_KEY}" "${CASSH_SERVER_URL}"/client)
+echo "${RESP}" > /tmp/test-cert
+OUTPUT=$(echo $(echo -n "$(ssh-keygen -L -f /tmp/test-cert 2>&1)"))
+if [[ "${OUTPUT}" == *"Principals: ${GUEST_B_USERNAME} Critical"* ]]; then
+    echo "[OK] Test signing key with updated principals"
+else
+    echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test signing key with updated principals: ${OUTPUT}"
+fi
+rm -f /tmp/test-cert
 
 ###################################
 ## ADMIN PURGE PRINCIPAL GUEST B ##
@@ -104,6 +100,16 @@ if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME}'"
 else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test purge principals to ${GUEST_B_USERNAME} : ${RESP}"
 fi
+
+RESP=$(curl -s -X POST -d "username=${GUEST_B_USERNAME}&realname=${GUEST_B_REALNAME}&password=${GUEST_B_PASSWORD}&pubkey=${GUEST_B_PUB_KEY}" "${CASSH_SERVER_URL}"/client)
+echo "${RESP}" > /tmp/test-cert
+OUTPUT=$(echo $(echo -n "$(ssh-keygen -L -f /tmp/test-cert 2>&1)"))
+if [[ "${OUTPUT}" == *"Principals: ${GUEST_B_USERNAME} Critical"* ]]; then
+    echo "[OK] Test signing key with updated principals"
+else
+    echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test signing key with updated principals: ${OUTPUT}"
+fi
+rm -f /tmp/test-cert
 
 RESP=$(curl -s -X POST -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "add=test-multiple-a,test-multiple-b")
 if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are '${GUEST_B_USERNAME},test-multiple-a,test-multiple-b'" ]; then
@@ -136,6 +142,16 @@ else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test update principals 'test-multiple-c,test-multiple-${GUEST_B_USERNAME}' to ${GUEST_B_USERNAME} : ${RESP}"
 fi
 
+RESP=$(curl -s -X POST -d "username=${GUEST_B_USERNAME}&realname=${GUEST_B_REALNAME}&password=${GUEST_B_PASSWORD}&pubkey=${GUEST_B_PUB_KEY}" "${CASSH_SERVER_URL}"/client)
+echo "${RESP}" > /tmp/test-cert
+OUTPUT=$(echo $(echo -n "$(ssh-keygen -L -f /tmp/test-cert 2>&1)"))
+if [[ "${OUTPUT}" == *"Principals: test-multiple-c test-multiple-${GUEST_B_USERNAME} Critical"* ]]; then
+    echo "[OK] Test signing key with updated principals"
+else
+    echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test signing key with updated principals: ${OUTPUT}"
+fi
+rm -f /tmp/test-cert
+
 RESP=$(curl -s -X POST -d "realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}" "${CASSH_SERVER_URL}"/admin/"${GUEST_B_USERNAME}"/principals -d "update=test-multiple-c,test-multiple-c,test-multiple-${GUEST_B_USERNAME}")
 if [ "${RESP}" == "OK: ${GUEST_B_USERNAME} principals are 'test-multiple-c,test-multiple-${GUEST_B_USERNAME}'" ]; then
     echo "[OK] Test update with duplicate principals 'test-multiple-c,test-multiple-${GUEST_B_USERNAME}' to ${GUEST_B_USERNAME}"
@@ -149,3 +165,13 @@ if [ "${RESP}" == "[ERROR] Unknown action" ]; then
 else
     echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test unknown action : ${RESP}"
 fi
+
+RESP=$(curl -s -X POST -d "username=${SYSADMIN_USERNAME}&realname=${SYSADMIN_REALNAME}&password=${SYSADMIN_PASSWORD}&pubkey=${SYSADMIN_PUB_KEY}" "${CASSH_SERVER_URL}"/client)
+echo "${RESP}" > /tmp/test-cert
+OUTPUT=$(echo $(echo -n "$(ssh-keygen -L -f /tmp/test-cert 2>&1)"))
+if [[ "${OUTPUT}" == *"Principals: ${SYSADMIN_USERNAME} Critical"* ]]; then
+    echo "[OK] Test signing sysadmin key"
+else
+    echo "[FAIL ${BASH_SOURCE}:+${LINENO}] Test signing sysadmin key: ${OUTPUT}"
+fi
+rm -f /tmp/test-cert